project/requirements.txt
alabaster==0.7.12
alembic==1.6.5
altair==4.1.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
astroid==2.6.5
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
base58==2.1.0
beautifulsoup4==4.9.3
black==20.8b1
bleach==3.3.0
blinker==1.4
Bootstrap-Flask==1.7.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
cfgv==3.3.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
distlib==0.3.2
Django==3.0.12
django-active-link==0.1.7
django-allauth==0.44.0
django-anymail==8.2
django-classy-tags==2.0.0
django-cors-headers==3.8.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.11.0
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.1
django-filter==2.4.0
django-gdpr-assist==1.3.0
django-localflavor==3.1
django-model-utils==4.1.1
django-parler==2.2
django-privacy-mgmt==0.0.21
django-redis==4.12.1
django-sekizai==2.0.0
django-stubs==1.7.0
django-stubs-ext==0.2.0
django-yaa-settings==1.1.0
djangorestframework==3.12.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
factory-boy==3.2.0
Faker==8.10.3
filelock==3.0.12
flake8==3.8.4
flake8-isort==4.0.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.2
furo==2021.4.11b34
gitdb==4.0.7
GitPython==3.1.18
greenlet==1.1.1
gunicorn==20.0.4
hiredis==1.1.0
identify==2.2.11
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
ipdb==0.13.4
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.2
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.6.0
livereload==2.6.3
Mako==1.1.4
Markdown==3.3.4
markdown2==2.4.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mypy==0.800
mypy-extensions==0.4.3
mysqlclient==2.0.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nodeenv==1.6.0
notebook==6.4.0
numpy==1.21.0
oauthlib==3.1.1
packaging==20.9
pandas==1.2.5
pandocfilters==1.4.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
platformdirs==2.1.0
plotly==4.9.0
pluggy==0.13.1
pre-commit==2.10.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.13.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydeck==0.6.2
pyflakes==2.2.0
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.9.5
pylint-django==2.4.2
pylint-plugin-utils==0.6
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytest-django==4.1.0
pytest-sugar==0.9.4
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
python-stdnum==1.16
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
redis==3.5.3
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
Send2Trash==1.7.1
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==3.4.3
sphinx-autobuild==2020.9.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.22
sqlparse==0.4.1
streamlit==0.83.0
stripe==2.60.0
termcolor==1.1.0
terminado==0.10.1
testfixtures==6.18.0
testpath==0.5.0
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.5
validators==0.18.2
virtualenv==20.6.0
watchdog==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
WTForms==2.3.3
